Moved slices inside feature directory
