+++ src/store.js
@@ -1,6 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit';
 import logger from 'redux-logger';
-import missionSlice from './slices/missionSlice';
+import missionSlice from './slices/missions/missionSlice';
 
 const store = configureStore({
   reducer: {
